Extract convertParametersToMbql and inline VARIABLE_ICONS
