+++ frontend/src/metabase-lib/v1/Question.ts
@@ -4,7 +4,6 @@ import { assoc, assocIn, chain, dissoc, getIn } from "icepick";
 import slugg from "slugg";
 import _ from "underscore";
 
-import { applyParameter } from "metabase/querying/parameters/utils/query";
 import * as Lib from "metabase-lib";
 import type Database from "metabase-lib/v1/metadata/Database";
 import Metadata from "metabase-lib/v1/metadata/Metadata";
@@ -29,7 +28,6 @@ import type {
   DatasetQuery,
   Field,
   LastEditInfo,
-  ParameterDimensionTarget,
   ParameterId,
   Parameter as ParameterObject,
   ParameterValuesMap,
@@ -38,7 +36,6 @@ import type {
   VisualizationDisplay,
   VisualizationSettings,
 } from "metabase-types/api";
-import { isDimensionTarget } from "metabase-types/guards";
 
 import type { Query } from "../query/types";
 
@@ -690,51 +687,6 @@ class Question {
     }
 
     return res;
-  }
-
-  _convertParametersToMbql({ isComposed }: { isComposed: boolean }): Question {
-    const query = this.query();
-    const { isNative } = Lib.queryDisplayInfo(query);
-
-    if (isNative) {
-      return this;
-    }
-
-    // If the query is composed (models or metrics) we cannot add filters to the underlying query since that query is used for data source.
-    // Pivot tables cannot work when there is an extra stage added on top of breakouts and aggregations.
-    const queryWithExtraStage =
-      !isComposed && this.display() !== "pivot"
-        ? Lib.ensureFilterStage(query)
-        : query;
-    const queryWithFilters = this.parameters().reduce((newQuery, parameter) => {
-      const stageIndex =
-        isDimensionTarget(parameter.target) && !isComposed
-          ? getParameterDimensionTargetStageIndex(parameter.target)
-          : -1;
-      return applyParameter(
-        newQuery,
-        stageIndex,
-        parameter.type,
-        parameter.target,
-        parameter.value,
-      );
-    }, queryWithExtraStage);
-    const queryWithFiltersWithoutExtraStage =
-      Lib.dropEmptyStages(queryWithFilters);
-
-    const newQuestion = this.setQuery(queryWithFiltersWithoutExtraStage)
-      .setParameters(undefined)
-      .setParameterValues(undefined);
-
-    const hasQueryBeenAltered = queryWithExtraStage !== queryWithFilters;
-    return hasQueryBeenAltered ? newQuestion.markDirty() : newQuestion;
-
-    function getParameterDimensionTargetStageIndex(
-      target: ParameterDimensionTarget,
-    ) {
-      const [_type, _variableTarget, options] = target;
-      return options?.["stage-number"] ?? -1;
-    }
   }
 
   query(): Query {
