+++ frontend/src/metabase-lib/v1/variables/TemplateTagVariable/TemplateTagVariable.ts
@@ -2,8 +2,6 @@ import NativeQuery from "metabase-lib/v1/queries/NativeQuery";
 import Variable from "metabase-lib/v1/variables/Variable";
 import type { TemplateTag, VariableTarget } from "metabase-types/api";
 
-import { VARIABLE_ICONS } from "./constants";
-
 // eslint-disable-next-line import/no-default-export -- deprecated usage
 export default class TemplateTagVariable extends Variable {
   tag(): TemplateTag | null | undefined {
@@ -15,11 +13,6 @@ export default class TemplateTagVariable extends Variable {
   displayName(): string | null | undefined {
     const tag = this.tag();
     return tag && (tag["display-name"] || tag.name);
-  }
-
-  icon(): string | null | undefined {
-    const tag = this.tag();
-    return tag && VARIABLE_ICONS[tag.type];
   }
 
   mbql(): VariableTarget {
